Добавление интерфейсов для классов сервиса
