+++ src/main/java/ru/yandex/practicum/filmorate/controller/DirectorController.java
@@ -2,6 +2,7 @@ package ru.yandex.practicum.filmorate.controller;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
+import java.util.Collection;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.validation.annotation.Validated;
@@ -14,10 +15,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import ru.yandex.practicum.filmorate.dto.DirectorDto;
-import ru.yandex.practicum.filmorate.model.Director;
 import ru.yandex.practicum.filmorate.service.DirectorService;
-
-import java.util.List;
 
 @Slf4j
 @Validated
@@ -26,35 +24,35 @@ import java.util.List;
 @RequiredArgsConstructor
 public class DirectorController {
 
-    private final DirectorService service;
+  private final DirectorService service;
 
-    @GetMapping
-    public List<DirectorDto> getAllDirectors() {
-        log.info("Received GET /directors");
-        return service.getAll();
-    }
+  @GetMapping
+  public Collection<DirectorDto> getAllDirectors() {
+    log.info("Received GET /directors");
+    return service.getAll();
+  }
 
-    @GetMapping("/{id}")
-    public DirectorDto getDirectorsById(@PathVariable("id") @NotNull @Positive final Long id) {
-        log.info("Received GET /directors/{}", id);
-        return service.getById(id);
-    }
+  @GetMapping("/{id}")
+  public DirectorDto getDirectorsById(@PathVariable("id") @NotNull @Positive final Long id) {
+    log.info("Received GET /directors/{}", id);
+    return service.getById(id);
+  }
 
-    @PostMapping
-    public DirectorDto save(@Validated @RequestBody final Director director) {
-        log.info("Received POST /directors - {}", director);
-        return service.save(director);
-    }
+  @PostMapping
+  public DirectorDto save(@Validated @RequestBody final DirectorDto director) {
+    log.info("Received POST /directors - {}", director);
+    return service.save(director);
+  }
 
-    @PutMapping
-    public DirectorDto update(@Validated @RequestBody final Director director) {
-        log.info("Received PUT /directors - {}", director);
-        return service.update(director);
-    }
+  @PutMapping
+  public DirectorDto update(@Validated @RequestBody final DirectorDto director) {
+    log.info("Received PUT /directors - {}", director);
+    return service.update(director);
+  }
 
-    @DeleteMapping("/{id}")
-    public void delete(@PathVariable("id") @NotNull @Positive final Long id) {
-        log.info("Received DELETE /directors/{}", id);
-        service.delete(id);
-    }
+  @DeleteMapping("/{id}")
+  public void delete(@PathVariable("id") @NotNull @Positive final Long id) {
+    log.info("Received DELETE /directors/{}", id);
+    service.removeById(id);
+  }
 }
