+++ src/main/java/ru/yandex/practicum/filmorate/mapper/DirectorMapper.java
@@ -6,10 +6,20 @@ import ru.yandex.practicum.filmorate.model.Director;
 
 @UtilityClass
 public class DirectorMapper {
-    public DirectorDto mapToDirectorDto(final Director director) {
-        return DirectorDto.builder()
-                .id(director.getId())
-                .name(director.getName())
-                .build();
-    }
+
+  public DirectorDto mapToDirectorDto(final Director director) {
+    return DirectorDto.builder()
+        .id(director.getId())
+        .name(director.getName())
+        .build();
+  }
+
+  public Director mapToDirector(final DirectorDto directorDto) {
+    return directorDto != null ?
+        Director.builder()
+            .id(directorDto.getId())
+            .name(directorDto.getName())
+            .build() :
+        null;
+  }
 }
